refactor: split the chainstore trait to use snapshot as an implementation detail
