+++ crates/amaru-consensus/src/stages/adopt_chain/test_setup.rs
@@ -19,7 +19,7 @@ use amaru_kernel::{
     cardano::network_block::make_encoded_block, make_header, make_header_with_op_cert_seq,
 };
 use amaru_ouroboros::{MempoolMsg, StoreError};
-use amaru_ouroboros_traits::{ChainStore, in_memory_consensus_store::InMemConsensusStore};
+use amaru_ouroboros_traits::{ChainStore, in_memory_chain_store::InMemoryChainStore};
 use amaru_protocols::store_effects::{
     FindAncestorOnBestChainEffect, FindAnchorAtHeightEffect, GetAnchorHashEffect, GetBestChainHashEffect,
     LoadFromBestChainEffect, LoadHeaderEffect, NextBestChainEffect, ResourceHeaderStore, RollForwardChainEffect,
@@ -165,7 +165,7 @@ pub fn test_prep(consensus_security_param: u64) -> TestPrep {
         state,
         rt: Builder::new_current_thread().build().unwrap(),
         headers,
-        store: Arc::new(InMemConsensusStore::new()),
+        store: Arc::new(InMemoryChainStore::new()),
     }
 }
