+++ crates/amaru-ouroboros-traits/src/stores/chain_store/diagnostic_chain_store.rs
@@ -1,0 +1,32 @@
+// Copyright 2026 PRAGMA
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+use amaru_kernel::{BlockHeader, HeaderHash, RawBlock};
+
+use crate::{Nonces, ReadChainStore};
+
+/// A chain store interface that exposes diagnostic methods to load raw data.
+/// It should not be used by the consensus stages since it might load lots of data at once.
+pub trait DiagnosticChainStore: ReadChainStore<BlockHeader> {
+    /// Load all headers in the store.
+    ///
+    /// NOTE: This can be very expensive for large stores and is only
+    /// used for diagnostics and testing purposes.
+    fn load_headers(&self) -> Box<dyn Iterator<Item = BlockHeader> + '_>;
+
+    /// Load all nonces in the store.
+    fn load_nonces(&self) -> Box<dyn Iterator<Item = (HeaderHash, Nonces)> + '_>;
+    fn load_blocks(&self) -> Box<dyn Iterator<Item = (HeaderHash, RawBlock)> + '_>;
+    fn load_parents_children(&self) -> Box<dyn Iterator<Item = (HeaderHash, Vec<HeaderHash>)> + '_>;
+}
