+++ crates/amaru-protocols/src/tx_submission/tests/system_test.rs
@@ -17,7 +17,7 @@ use std::{sync::Arc, time::Duration};
 use amaru_kernel::{EraHistory, NetworkMagic, NetworkName, Peer, Transaction};
 use amaru_mempool::InMemoryMempool;
 use amaru_network::connection::TokioConnections;
-use amaru_ouroboros::{ConnectionsResource, in_memory_consensus_store::InMemConsensusStore};
+use amaru_ouroboros::{ConnectionsResource, in_memory_chain_store::InMemoryChainStore};
 use amaru_ouroboros_traits::ResourceMempool;
 use pure_stage::{StageGraph, StageRef, tokio::TokioBuilder};
 use tokio::{runtime::Handle, time::timeout};
@@ -48,7 +48,7 @@ async fn test_tx_submission_with_node() -> anyhow::Result<()> {
     let mut network = TokioBuilder::default();
 
     network.resources().put::<ConnectionsResource>(Arc::new(conn));
-    network.resources().put::<ResourceHeaderStore>(Arc::new(InMemConsensusStore::new()));
+    network.resources().put::<ResourceHeaderStore>(Arc::new(InMemoryChainStore::new()));
     network.resources().put::<ResourceMempool<Transaction>>(Arc::new(InMemoryMempool::default()));
 
     let era_history: &EraHistory = NetworkName::Preprod.into();
