+++ crates/amaru/src/bootstrap.rs
@@ -28,7 +28,7 @@ use amaru_ledger::{
     store::{EpochTransitionProgress, Store, TransactionalContext},
 };
 use amaru_network::chain_sync_client::ChainSyncClient;
-use amaru_ouroboros::{ChainStore, Nonces};
+use amaru_ouroboros::{ChainStore, Nonces, WriteChainStore};
 use amaru_progress_bar::{ProgressBar, TerminalProgressBar};
 use amaru_stores::rocksdb::{RocksDB, RocksDbConfig, consensus::RocksDBStore};
 use async_compression::tokio::bufread::GzipDecoder as AsyncGzipDecoder;
